Add files via upload
Brief Introduction: In this assignment, we are helping our client company, which buys used cars from sellers in the area. We will be building machine learning models that could predict the prices of used cars in order to speed up their process. This will replace their traditional method of evaluating used car prices. Our models will take into account the car features and quote a price for the seller on behalf of our company.